server/requirements.txt
numpy
sounddevice 
vosk 
openai-whisper